refactor : StartGame 내의 함수 분리
