+++ NumberBaseball/NumberBaseball/Extensions/Array.swift
@@ -1,0 +1,5 @@
+extension Array where Element == Int {
+	func generateDescription() -> String {
+		return self.map { String($0) }.joined(separator: " ")
+	}
+}
